AG-13008 Ignore KeyNavEvent<'delete'> when text input is shown

+++ packages/ag-charts-enterprise/src/features/annotations/annotations.ts
@@ -1227,6 +1227,7 @@ export class Annotations extends _ModuleSupport.BaseModuleInstance implements _M
     }
 
     private onDelete() {
+        if (this.textInput.isVisible()) return;
         this.cancel();
         this.delete();
         this.reset();
